fix: move achievementcategory display to separate component
