+++ src/routes/achievements/+page.svelte
@@ -1,27 +1,11 @@
 <script lang="ts">
 	import { applicationState } from '$lib/store/application_state';
-	import { get, derived } from 'svelte/store';
+	import { get } from 'svelte/store';
 	import { userProfile } from '$lib/store/user_profile';
 	import DefaultLayout from '$lib/components/layout/DefaultLayout.svelte';
-	import { dataIndexStore } from '$lib/store/index_store';
 	import i18n from '$lib/services/i18n';
-	import type { AchievementCategoryKey } from '$lib/types/keys/AchievementCategoryKey';
-	import AchievementCategoryCard from '$lib/components/ui/card/AchievementCategoryCard.svelte';
-	import S3Service from '$lib/services/s3';
+	import AchievementCategoryDisplay from '$lib/components/navigator/AchievementCategoryDisplay.svelte';
 
-	const transformedAchievements = derived([dataIndexStore], ([dataIndexStore]) => {
-		const achievements = Object.keys(dataIndexStore.achievementCategory).map((key) => {
-			const achievement = dataIndexStore.achievementCategory[key as AchievementCategoryKey];
-			return {
-				link: `/achievements/${key}`,
-				name: achievement!.name,
-				order: achievement!.order,
-				img: S3Service.getAchievementCategory(key as AchievementCategoryKey).icon,
-				total: achievement!.total
-			};
-		});
-		return achievements.sort((a, b) => a.order - b.order);
-	});
 	console.log(get(applicationState));
 	console.log(get(userProfile));
 </script>
@@ -30,15 +14,5 @@
 <!-- svelte-ignore a11y-label-has-associated-control -->
 <DefaultLayout title={$i18n.t('achievements.categories.title')}>
 	<svelte:fragment slot="titlebarActions"></svelte:fragment>
-	<div class="flex flex-wrap gap-3 justify-center">
-		{#each $transformedAchievements as achievement}
-			<AchievementCategoryCard
-				link={achievement.link}
-				name={achievement.name}
-				img={achievement.img}
-				total={achievement.total}
-				achieved={achievement.total}
-			/>
-		{/each}
-	</div>
+	<AchievementCategoryDisplay></AchievementCategoryDisplay>
 </DefaultLayout>
