+++ src/routes/achievements/[slug]/+page.svelte
@@ -4,7 +4,6 @@
 	import DefaultLayout from '$lib/components/layout/DefaultLayout.svelte';
 	import Card from '../../../lib/components/ui/card/card.svelte';
 	import Text from '$lib/components/typography/Text.svelte';
-
 	export let data: PageData;
 
 	const fetchAchievementCategoryData = data.backend.data.fetchAchievementCategoryData(
@@ -14,11 +13,9 @@
 
 <DefaultLayout isLoading={$fetchAchievementCategoryData.status === 'pending'}>
 	{#if $fetchAchievementCategoryData.status === 'success'}
-		<section class="grid grid-cols-1 gap-4">
+		<div class="grid grid-cols-1 gap-4">
 			{#each $fetchAchievementCategoryData.data.achievements as achievement}
-				<Card
-					class="flex flex-col bg-neutral border-secondary h-full w-full lg:overflow-auto sm:p-0 p-3"
-				>
+				<Card class="flex flex-col bg-neutral h-full w-full lg:overflow-auto sm:p-0 p-3">
 					<Text type="h4" class="font-bold text-lg w-full px-2 py-0">
 						{achievement.name}
 					</Text>
@@ -27,6 +24,6 @@
 					</Text>
 				</Card>
 			{/each}
-		</section>
+		</div>
 	{/if}
 </DefaultLayout>
